+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### New Features
 - Use Java runtimes (when available) for Minecraft [#459]
+- Add in new Vanilla pack installer + support for snapshots [#460]
 
 ### Fixes
 - Launcher not working with Java 16 [#465]